style: rendering minor code cleanup

+++ src/Runtime/include/NeonRuntime/Backends/OpenGLRenderer.hpp
@@ -23,10 +23,10 @@
 namespace Neon
 {
     struct Character {
-        unsigned int TextureID;  // ID handle of the glyph texture
+        signed long TextureID;  // ID handle of the glyph texture
         glm::ivec2   Size;       // Size of glyph
         glm::ivec2   Bearing;    // Offset from baseline to left/top of glyph
-        unsigned int Advance;    // Offset to advance to next glyph
+        signed long Advance;    // Offset to advance to next glyph
     };
 
     class OpenGLRenderer final : public IRenderer
